Add --run-name option
It is usefull to have possibility to assign custom run name.
First of all, if teuthology run is scheduled from external
continuous integration system it makes it easy to refer
to corresponding job.

Second, it solves the problem when several different runs
are scheduled at the same time, which allows to avoid jobs
get into different run.